misc: Rename stencil transformations files (#1134)
Just cleaning up file names.
