+++ xdsl/transforms/experimental/stencil_storage_materialization.py
